requirements.txt
langchain
langchain-openai
langchain-community
langgraph
tavily-python
streamlit
python-dotenv